api/requirements.txt
Flask
Flask-CORS
openai
boto3
PyYAML
pydantic
python-dotenv
filelock
gunicorn